WIP replace test UI with real UI in tags feed screen
