+++ WordPress/src/main/java/org/wordpress/android/ui/reader/views/compose/tagsfeed/ReaderTagsFeed.kt
@@ -48,6 +48,11 @@ import org.wordpress.android.models.ReaderTagType
 import org.wordpress.android.ui.compose.theme.AppColor
 import org.wordpress.android.ui.compose.theme.AppTheme
 import org.wordpress.android.ui.compose.unit.Margin
+import org.wordpress.android.ui.reader.viewmodels.ReaderTagsFeedViewModel.ErrorType
+import org.wordpress.android.ui.reader.viewmodels.ReaderTagsFeedViewModel.PostList
+import org.wordpress.android.ui.reader.viewmodels.ReaderTagsFeedViewModel.TagChip
+import org.wordpress.android.ui.reader.viewmodels.ReaderTagsFeedViewModel.TagFeedItem
+import org.wordpress.android.ui.reader.viewmodels.ReaderTagsFeedViewModel.UiState
 import org.wordpress.android.ui.reader.views.compose.filter.ReaderFilterChip
 import org.wordpress.android.ui.utils.UiString
 import org.wordpress.android.util.AppLog
@@ -63,6 +68,9 @@ fun ReaderTagsFeed(uiState: UiState) {
             is UiState.Loading -> Loading()
             is UiState.Loaded -> Loaded(uiState)
             is UiState.Empty -> Empty(uiState)
+            is UiState.Initial -> {
+                // no-op
+            }
         }
     }
 }
@@ -405,42 +413,6 @@ private fun PostListError(
             )
         }
     }
-}
-
-// TODO move to VM
-sealed class UiState {
-    data class Loaded(val data: List<TagFeedItem>) : UiState()
-
-    object Loading : UiState()
-
-    data class Empty(val onOpenTagsListClick: () -> Unit) : UiState()
-}
-
-data class TagFeedItem(
-    val tagChip: TagChip,
-    val postList: PostList,
-)
-
-data class TagChip(
-    val tag: ReaderTag,
-    val onTagClicked: () -> Unit,
-)
-
-sealed class PostList {
-    data class Loaded(val items: List<TagsFeedPostItem>) : PostList()
-
-    object Loading : PostList()
-
-    data class Error(
-        val type: ErrorType,
-        val onRetryClick: () -> Unit
-    ) : PostList()
-}
-
-sealed interface ErrorType {
-    data object Loading : ErrorType
-
-    data object NoContent : ErrorType
 }
 
 data class TagsFeedPostItem(
